update external code

+++ DalamudPlugin/RaidsRewritten/Structures/CachedObjectEffectInfo.cs
@@ -1,5 +1,5 @@
 ﻿// Adapted from https://github.com/PunishXIV/Splatoon/blob/main/Splatoon/Structures/CachedObjectEffectInfo.cs
-// 0054cc3
+// 534f6cb
 using System;
 
 namespace RaidsRewritten.Structures;
@@ -7,10 +7,10 @@ namespace RaidsRewritten.Structures;
 public record struct CachedObjectEffectInfo
 {
     public long StartTime;
-    public ushort data1;
-    public ushort data2;
+    public uint data1;
+    public uint data2;
 
-    public CachedObjectEffectInfo(long startTime, ushort data1, ushort data2)
+    public CachedObjectEffectInfo(long startTime, uint data1, uint data2)
     {
         StartTime = startTime;
         this.data1 = data1;
